move files into cmd and resolvers directories
