+++ resolvers/http_resolver/http_resolver_test.go
