+++ resolvers/targets/targets.go
@@ -1,10 +1,10 @@
 package targets
 
 import (
-	"github.com/kitsuyui/myip/base"
-	"github.com/kitsuyui/myip/dns_resolver"
-	"github.com/kitsuyui/myip/http_resolver"
-	"github.com/kitsuyui/myip/stun_resolver"
+	"github.com/kitsuyui/myip/resolvers/base"
+	"github.com/kitsuyui/myip/resolvers/dns_resolver"
+	"github.com/kitsuyui/myip/resolvers/http_resolver"
+	"github.com/kitsuyui/myip/resolvers/stun_resolver"
 )
 
 func IPRetrievables() []base.ScoredIPRetrievable {
